+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v55.0.0-SNAPSHOT - unreleased
 
+### 💥 Breaking Changes
+
+* The use of `tooltipElement` on `Column` is deprecated. Use `tooltip` instead.
+
 ## v54.0.0 - 2022-12-31
 
 We are pleased to announce that Hoist React has been fully rewritten in TypeScript! ✨🚀